+++ config/config_test.go
@@ -22,8 +22,24 @@ func TestMain(m *testing.M) {
 	os.Exit(m.Run())
 }
 
+func testFile(t *testing.T, contents string) (path string, remove func()) {
+	testFile, err := ioutil.TempFile("", "test.")
+	if err != nil {
+		t.Fatal(err)
+	}
+	if len(contents) > 0 {
+		if _, err := testFile.Write([]byte(contents)); err != nil {
+			t.Fatal(err)
+		}
+	}
+	return testFile.Name(), func() { os.Remove(testFile.Name()) }
+}
+
 func TestParse(t *testing.T) {
 	t.Parallel()
+
+	testFilePath, remove := testFile(t, "")
+	defer remove()
 
 	cases := []struct {
 		name string
@@ -384,16 +400,16 @@ func TestParse(t *testing.T) {
 			false,
 		},
 		{
-			"exec_env_blacklist",
+			"exec_env_denylist",
 			`exec {
 				env {
-					blacklist = ["a", "b"]
+					denylist = ["a", "b"]
 				}
 			 }`,
 			&Config{
 				Exec: &ExecConfig{
 					Env: &EnvConfig{
-						Blacklist: []string{"a", "b"},
+						Denylist: []string{"a", "b"},
 					},
 				},
 			},
@@ -432,16 +448,16 @@ func TestParse(t *testing.T) {
 			false,
 		},
 		{
-			"exec_env_whitelist",
+			"exec_env_allowlist",
 			`exec {
 				env {
-					whitelist = ["a", "b"]
+					allowlist = ["a", "b"]
 				}
 			 }`,
 			&Config{
 				Exec: &ExecConfig{
 					Env: &EnvConfig{
-						Whitelist: []string{"a", "b"},
+						Allowlist: []string{"a", "b"},
 					},
 				},
 			},
@@ -533,9 +549,9 @@ func TestParse(t *testing.T) {
 		},
 		{
 			"block_query_wait",
-			`block_query_wait = "60s"`,
+			`block_query_wait = "61s"`,
 			&Config{
-				BlockQueryWaitTime: TimeDuration(60 * time.Second),
+				BlockQueryWaitTime: TimeDuration(61 * time.Second),
 			},
 			false,
 		},
@@ -774,7 +790,29 @@ func TestParse(t *testing.T) {
 			false,
 		},
 		{
-			"template_exec_env_blacklist",
+			"template_exec_env_denylist",
+			`template {
+				exec {
+					env {
+						denylist = ["a", "b"]
+					}
+				}
+			 }`,
+			&Config{
+				Templates: &TemplateConfigs{
+					&TemplateConfig{
+						Exec: &ExecConfig{
+							Env: &EnvConfig{
+								Denylist: []string{"a", "b"},
+							},
+						},
+					},
+				},
+			},
+			false,
+		},
+		{
+			"template_exec_env_denylist_deprecated",
 			`template {
 				exec {
 					env {
@@ -787,7 +825,7 @@ func TestParse(t *testing.T) {
 					&TemplateConfig{
 						Exec: &ExecConfig{
 							Env: &EnvConfig{
-								Blacklist: []string{"a", "b"},
+								DenylistDeprecated: []string{"a", "b"},
 							},
 						},
 					},
@@ -840,7 +878,29 @@ func TestParse(t *testing.T) {
 			false,
 		},
 		{
-			"template_exec_env_whitelist",
+			"template_exec_env_allowlist",
+			`template {
+				exec {
+					env {
+						allowlist = ["a", "b"]
+					}
+				}
+			 }`,
+			&Config{
+				Templates: &TemplateConfigs{
+					&TemplateConfig{
+						Exec: &ExecConfig{
+							Env: &EnvConfig{
+								Allowlist: []string{"a", "b"},
+							},
+						},
+					},
+				},
+			},
+			false,
+		},
+		{
+			"template_exec_env_allowlist_deprecated",
 			`template {
 				exec {
 					env {
@@ -853,7 +913,7 @@ func TestParse(t *testing.T) {
 					&TemplateConfig{
 						Exec: &ExecConfig{
 							Env: &EnvConfig{
-								Whitelist: []string{"a", "b"},
+								AllowlistDeprecated: []string{"a", "b"},
 							},
 						},
 					},
@@ -1085,6 +1145,18 @@ func TestParse(t *testing.T) {
 			&Config{
 				Vault: &VaultConfig{
 					Token: String("token"),
+				},
+			},
+			false,
+		},
+		{
+			"vault_agent_token_file",
+			`vault {
+				vault_agent_token_file = "` + testFilePath + `"
+			}`,
+			&Config{
+				Vault: &VaultConfig{
+					VaultAgentTokenFile: String(testFilePath),
 				},
 			},
 			false,
@@ -1489,14 +1561,24 @@ func TestParse(t *testing.T) {
 func TestFinalize(t *testing.T) {
 	t.Parallel()
 
+	testFileContents := "testing123"
+	testFilePath, remove := testFile(t, testFileContents)
+	defer remove()
+
+	type finalizer interface {
+		Finalize()
+	}
+
 	// Testing Once disabling Wait
 	cases := []struct {
-		name string
-		expt *Config
-		test *Config
+		name    string
+		isEqual func(*Config, *Config) (bool, error)
+		test    *Config
+		expt    *Config
 	}{
 		{
 			"null-case",
+			nil,
 			&Config{
 				Wait: &WaitConfig{
 					Enabled: Bool(true),
@@ -1510,18 +1592,42 @@ func TestFinalize(t *testing.T) {
 					Min:     TimeDuration(10 * time.Second),
 					Max:     TimeDuration(20 * time.Second),
 				},
-				Once: false,
+			},
+		},
+		{
+			"vault_agent_token_file",
+			func(act, exp *Config) (bool, error) {
+				actToken := *act.Vault.Token
+				expToken := *exp.Vault.Token
+				if actToken != expToken {
+					return false, fmt.Errorf("tokens don't match: %v != %v",
+						actToken, expToken)
+				}
+				actTokenFile := *act.Vault.VaultAgentTokenFile
+				expTokenFile := *exp.Vault.VaultAgentTokenFile
+				if actTokenFile != expTokenFile {
+					return false, fmt.Errorf("tokenfiles don't match: %v != %v",
+						actTokenFile, expTokenFile)
+				}
+				return true, nil
+			},
+			&Config{
+				Vault: &VaultConfig{
+					VaultAgentTokenFile: String(testFilePath),
+				},
+			},
+			&Config{
+				Vault: &VaultConfig{
+					Address:             String(""),
+					Namespace:           String(""),
+					VaultAgentTokenFile: String(testFilePath),
+					Token:               String(testFileContents),
+				},
 			},
 		},
 		{
 			"once-disables-wait",
-			&Config{
-				Wait: &WaitConfig{
-					Enabled: Bool(false),
-					Min:     nil,
-					Max:     nil,
-				},
-			},
+			nil,
 			&Config{
 				Wait: &WaitConfig{
 					Enabled: Bool(true),
@@ -1530,14 +1636,28 @@ func TestFinalize(t *testing.T) {
 				},
 				Once: true,
 			},
+			&Config{
+				Wait: &WaitConfig{
+					Enabled: Bool(false),
+					Min:     nil,
+					Max:     nil,
+				},
+			},
 		},
 	}
 
 	for i, tc := range cases {
 		t.Run(fmt.Sprintf("%d_%s", i, tc.name), func(t *testing.T) {
 			tc.test.Finalize()
-			if !reflect.DeepEqual(tc.expt.Wait, tc.test.Wait) {
-				t.Errorf("\nexp: %#v\nact: %#v", tc.expt.Wait, tc.test.Wait)
+			switch tc.isEqual {
+			case nil:
+				if !reflect.DeepEqual(tc.expt.Wait, tc.test.Wait) {
+					t.Errorf("\nexp: %#v\nact: %#v", *tc.expt.Wait, *tc.test.Wait)
+				}
+			default:
+				if eq, err := tc.isEqual(tc.test, tc.expt); !eq {
+					t.Errorf(err.Error())
+				}
 			}
 		})
 	}
@@ -1672,10 +1792,22 @@ func TestConfig_Merge(t *testing.T) {
 				BlockQueryWaitTime: TimeDuration(1 * time.Second),
 			},
 			&Config{
-				BlockQueryWaitTime: TimeDuration(60 * time.Second),
+				BlockQueryWaitTime: TimeDuration(61 * time.Second),
 			},
 			&Config{
-				BlockQueryWaitTime: TimeDuration(60 * time.Second),
+				BlockQueryWaitTime: TimeDuration(61 * time.Second),
+			},
+		},
+		{
+			"block_query_wait_nil",
+			&Config{
+				BlockQueryWaitTime: TimeDuration(1 * time.Second),
+			},
+			&Config{
+				BlockQueryWaitTime: nil,
+			},
+			&Config{
+				BlockQueryWaitTime: TimeDuration(1 * time.Second),
 			},
 		},
 		{