+++ dependency/vault_read_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/hashicorp/vault/api"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestNewVaultReadQuery(t *testing.T) {
@@ -398,6 +399,18 @@ func TestVaultReadQuery_Fetch_KVv2(t *testing.T) {
 		})
 	}
 
+	t.Run("read_metadata", func(t *testing.T) {
+		d, err := NewVaultReadQuery(secretsPath + "/metadata/foo/bar")
+		require.NoError(t, err)
+
+		act, _, err := d.Fetch(clients, nil)
+		require.NoError(t, err)
+		require.NotNil(t, act)
+
+		versions := act.(*Secret).Data["versions"]
+		assert.Len(t, versions, 2)
+	})
+
 	t.Run("read_deleted", func(t *testing.T) {
 		// only needed for KVv2 as KVv1 doesn't have metadata
 		path := "data/foo/zed"
@@ -648,6 +661,73 @@ func TestVaultReadQuery_String(t *testing.T) {
 				t.Fatal(err)
 			}
 			assert.Equal(t, tc.exp, d.String())
+		})
+	}
+}
+
+func TestShimKVv2Path(t *testing.T) {
+	cases := []struct {
+		name      string
+		path      string
+		mountPath string
+		expected  string
+	}{
+		{
+			"full path",
+			"secret/data/foo/bar",
+			"secret/",
+			"secret/data/foo/bar",
+		}, {
+			"data prefix added",
+			"secret/foo/bar",
+			"secret/",
+			"secret/data/foo/bar",
+		}, {
+			"full path with data* in subpath",
+			"secret/data/datafoo/bar",
+			"secret/",
+			"secret/data/datafoo/bar",
+		}, {
+			"prefix added with data* in subpath",
+			"secret/datafoo/bar",
+			"secret/",
+			"secret/data/datafoo/bar",
+		}, {
+			"prefix added with *data in subpath",
+			"secret/foodata/foo/bar",
+			"secret/",
+			"secret/data/foodata/foo/bar",
+		}, {
+			"prefix not added to metadata",
+			"secret/metadata/foo/bar",
+			"secret/",
+			"secret/metadata/foo/bar",
+		}, {
+			"prefix added with metadata* in subpath",
+			"secret/metadatafoo/foo/bar",
+			"secret/",
+			"secret/data/metadatafoo/foo/bar",
+		}, {
+			"prefix added with *metadata in subpath",
+			"secret/foometadata/foo/bar",
+			"secret/",
+			"secret/data/foometadata/foo/bar",
+		}, {
+			"prefix added to mount path",
+			"secret/",
+			"secret/",
+			"secret/data",
+		}, {
+			"prefix added to mount path not exact match",
+			"secret",
+			"secret/",
+			"secret/data",
+		},
+	}
+	for _, tc := range cases {
+		t.Run(tc.name, func(t *testing.T) {
+			actual := shimKVv2Path(tc.path, tc.mountPath)
+			assert.Equal(t, tc.expected, actual)
 		})
 	}
 }