+++ dependency/vault_token_test.go
@@ -49,8 +49,7 @@ func TestNewVaultTokenQuery(t *testing.T) {
 				act.stopCh = nil
 			}
 
-			assert.Equal(t, tc.exp.secret, act.secret)
-			assert.Equal(t, tc.exp.vaultSecret, act.vaultSecret)
+			assert.Equal(t, tc.exp, act)
 		})
 	}
 }