+++ main_test.go
@@ -8,6 +8,7 @@ import (
 	"testing"
 
 	dep "github.com/hashicorp/consul-template/dependency"
+	"github.com/hashicorp/consul-template/test"
 	"github.com/hashicorp/consul/sdk/testutil"
 )
 
@@ -16,11 +17,13 @@ var testClients *dep.ClientSet
 
 func TestMain(m *testing.M) {
 	log.SetOutput(ioutil.Discard)
-	consul, err := testutil.NewTestServerConfig(func(c *testutil.TestServerConfig) {
-		c.LogLevel = "warn"
-		c.Stdout = ioutil.Discard
-		c.Stderr = ioutil.Discard
-	})
+	tb := &test.TestingTB{}
+	consul, err := testutil.NewTestServerConfigT(tb,
+		func(c *testutil.TestServerConfig) {
+			c.LogLevel = "warn"
+			c.Stdout = ioutil.Discard
+			c.Stderr = ioutil.Discard
+		})
 	if err != nil {
 		log.Fatal(fmt.Errorf("failed to start consul server: %v", err))
 	}
@@ -52,6 +55,7 @@ func TestMain(m *testing.M) {
 
 	exit := <-exitCh
 
+	tb.DoCleanup()
 	testConsul.Stop()
 	os.Exit(exit)
 }