+++ template/template_test.go
@@ -1371,6 +1371,17 @@ func TestTemplate_Execute(t *testing.T) {
 			false,
 		},
 		{
+			"helper_parseFloat_format",
+			&NewTemplateInput{
+				Contents: `{{ "1.0" | parseFloat | printf "%.1f"}}`,
+			},
+			&ExecuteInput{
+				Brain: NewBrain(),
+			},
+			"1.0",
+			false,
+		},
+		{
 			"helper_parseInt",
 			&NewTemplateInput{
 				Contents: `{{ "-1" | parseInt }}`,
@@ -1450,8 +1461,8 @@ func TestTemplate_Execute(t *testing.T) {
 		{
 			"helper_plugin_disabled",
 			&NewTemplateInput{
-				Contents:          `{{ "1" | plugin "echo" }}`,
-				FunctionBlacklist: []string{"plugin"},
+				Contents:         `{{ "1" | plugin "echo" }}`,
+				FunctionDenylist: []string{"plugin"},
 			},
 			&ExecuteInput{
 				Brain: NewBrain(),
@@ -1534,6 +1545,44 @@ func TestTemplate_Execute(t *testing.T) {
 				Brain: NewBrain(),
 			},
 			"[\"a\",\"b\",\"c\"]",
+			false,
+		},
+		{
+			"helper_toUnescapedJSON",
+			&NewTemplateInput{
+				Contents: `{{ "a?b&c,x?y&z" | split "," | toUnescapedJSON }}`,
+			},
+			&ExecuteInput{
+				Brain: NewBrain(),
+			},
+			"[\"a?b&c\",\"x?y&z\"]",
+			false,
+		},
+		{
+			"helper_toUnescapedJSONPretty",
+			&NewTemplateInput{
+				Contents: `{{ tree "key" | explode | toUnescapedJSONPretty }}`,
+			},
+			&ExecuteInput{
+				Brain: func() *Brain {
+					b := NewBrain()
+					d, err := dep.NewKVListQuery("key")
+					if err != nil {
+						t.Fatal(err)
+					}
+					b.Remember(d, []*dep.KeyPair{
+						&dep.KeyPair{Key: "a", Value: "b&c"},
+						&dep.KeyPair{Key: "x", Value: "y&z"},
+						&dep.KeyPair{Key: "k", Value: "<>&&"},
+					})
+					return b
+				}(),
+			},
+			`{
+  "a": "b&c",
+  "k": "<>&&",
+  "x": "y&z"
+}`,
 			false,
 		},
 		{